requirements.txt
asgiref==3.11.1
Django==5.2.11
djangorestframework==3.16.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
